Moved and simplified LAFFResult
