+++ src/main/java/com/github/skjolberg/packing/LargestAreaFitFirstPackager.java
@@ -1,9 +1,6 @@
 package com.github.skjolberg.packing;
 
-import com.github.skjolberg.packing.impl.Adapter;
-import com.github.skjolberg.packing.impl.PackResult;
-import com.github.skjolberg.packing.impl.Placement;
-import com.github.skjolberg.packing.impl.Space;
+import com.github.skjolberg.packing.impl.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,57 +11,6 @@ import java.util.List;
  * Thread-safe implementation. The input Boxes must however only be used in a single thread at a time.
  */
 public class LargestAreaFitFirstPackager extends Packager {
-
-	private static class LAFFResult implements PackResult {
-
-		private List<Box> remaining;
-		private Container container;
-
-		public LAFFResult(List<Box> remaining, Container container) {
-			this.remaining = remaining;
-			this.container = container;
-		}
-
-		public Container getContainer() {
-			return container;
-		}
-
-		@Override
-		public boolean packsMoreBoxesThan(PackResult result) {
-			LAFFResult laffResult = (LAFFResult)result;
-			if(laffResult.remaining.size() > remaining.size()) { // lower is better
-				return true;
-			} else if(laffResult.remaining.size() > remaining.size()) {
-				// check volume (of container)
-				if(laffResult.container.getVolume() > container.getVolume()) {
-					return true;
-				} else if(laffResult.container.getVolume() == container.getVolume()) {
-					// check weight (of boxes)
-					if(laffResult.container.getStackWeight() > container.getStackWeight()) {
-						return true;
-					}
-				}
-			}
-
-			return false;
-		}
-
-		public List<Box> getRemainingBoxes() {
-			return remaining;
-		}
-
-		@Override
-		public boolean isEmpty() {
-			return container.getLevels().isEmpty() || container.getLevels().get(0).isEmpty();
-		}
-
-		@Override
-		public String toString() {
-			return "LAFFResult [container=" + container + "]";
-		}
-
-
-	}
 
 	protected final boolean footprintFirst;
